.requirements.txt
pytest
python-jsonschema-objects>=0.4.0
jsonschema==3.2.0
ipython
pyyaml
ga4gh.gks.metaschema==0.3.0b2
sphinx ~= 4.5
sphinx-rtd-theme ~= 1.2